backend/requirements.txt
Flask==3.1.0
flask-cors==5.0.0
duckdb==1.1.3
pandas==2.2.3
python-dotenv==1.0.0
google-generativeai==0.8.3
cloudinary==1.33.0
gunicorn==23.0.0
pymongo==4.13.2
PyJWT==2.10.1
Werkzeug==3.1.3
PyMySQL==1.1.0